Add GitHub Pages deployment workflow and fix lint issues
