+++ README.md
@@ -1,0 +1,15 @@
+# neilthomass.github.io
+
+Personal website built with Vite, React, and Tailwind CSS.
+
+## Development
+
+```bash
+npm install
+npm run dev
+```
+
+## Deployment
+
+The site is published to GitHub Pages via the workflow in
+[.github/workflows/deploy.yml](.github/workflows/deploy.yml).
